requirements.txt
torch
numpy
pandas
scikit-learn
matplotlib
streamlit
Pillow
plotly
